benchplot: initial commit
This is a tool for generating plots of Go benchmark results.

Currently it's fairly limited, but supports configuring the X and Y
axes, color, and 2-D faceting. It supports Gnuplot line charts with
continuous axes, but is designed to be easily extensible to other
types of plots and other plotting tools.