plot better/worse under the curve rather than shading the whole plot
Currently, when plotting ratio data, we color the entire area of the
plot above zero red or green and the entire area below the opposite.
However, visually this is misleading because a single point that's,
say, in the worse direction, can result in a large red area even if
most of the curve in the green area.

Replace this with coloring just between the curve and the "+0%" line.
This highlights the "area" under the curve, so you see more green if
more of the points are in the better direction.